Fix consistency of PATH with PERL5LIB
Fix the consistency of $ENV{PATH} with $ENV{PERL5LIB}.

This fix can be tested by "test"-ing CPANPLUS-Test-Dummy-UsesScript.

Some packages fail to install because the CPAN client has modified
$ENV{PERL5LIB} to add some temporary lib paths, but did not modify PATH
to add the bin or script dirs. This cause problems for distributions
that rely on Perl scripts instead of modules, as we can get trouble if
there is a mismatch between the script and its supporting modules
(example: running an older script with the newer modules, or finding the
module but not the newer script).

Here is an example issue:
https://rt.cpan.org/Ticket/Display.html?id=75376